# openapi/v2.json
{
  "components": {
    "schemas": {
...
      "ApplyRequest": {
        "additionalProperties": false,
        "properties": {
          "selections": {
            "items": {
              "$ref": "#/components/schemas/Selection"
            },
            "title": "Selections",
            "type": "array"
          },
          "snapshot_id": {
            "title": "Snapshot Id",
            "type": "string"
          }
        },
        "required": [
          "snapshot_id"
        ],
        "title": "ApplyRequest",
        "type": "object"
      },
...
      "CleanupResponse": {
        "additionalProperties": false,
        "properties": {
          "deleted": {
            "minimum": 0.0,
            "title": "Deleted",
            "type": "integer"
          },
          "examined": {
            "minimum": 0.0,
            "title": "Examined",
            "type": "integer"
          },
          "failed": {
            "minimum": 0.0,
            "title": "Failed",
            "type": "integer"
          }
        },
        "required": [
          "examined",
          "deleted",
          "failed"
        ],
        "title": "CleanupResponse",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "default": false,
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "SessionResponse": {
        "additionalProperties": false,
        "properties": {
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "session_token": {
            "title": "Session Token",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "session_token",
          "expires_at"
        ],
        "title": "SessionResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "SessionBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "TTML ID Match API",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v2/maintenance/cleanup": {
      "get": {
        "operationId": "cleanup_expired_sessions_api_v2_maintenance_cleanup_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CleanupResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          }
        },
        "summary": "Cleanup Expired Sessions"
      }
    },
    "/api/v2/sessions": {
      "post": {
        "operationId": "create_session_api_v2_sessions_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          }
        },
        "summary": "Create Session"
      }
    },
    "/api/v2/sessions/{session_id}": {
      "delete": {
        "operationId": "delete_session_api_v2_sessions__session_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          }
        },
        "security": [
          {
            "SessionBearer": []
          }
        ],
        "summary": "Delete Session"
      }
    },
    "/api/v2/sessions/{session_id}/apply": {
      "post": {
        "operationId": "apply_api_v2_sessions__session_id__apply_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          }
        },
        "security": [
          {
            "SessionBearer": []
          }
        ],
        "summary": "Apply"
      }
    },
    "/api/v2/sessions/{session_id}/change-plans": {
      "post": {
        "operationId": "change_plan_api_v2_sessions__session_id__change_plans_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangePlanResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          }
        },
        "security": [
          {
            "SessionBearer": []
          }
        ],
        "summary": "Change Plan"
      }
    },
    "/api/v2/sessions/{session_id}/files": {
      "post": {
        "operationId": "upload_files_api_v2_sessions__session_id__files_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_files_api_v2_sessions__session_id__files_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairingPlanResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          }
        },
        "security": [
          {
            "SessionBearer": []
          }
        ],
        "summary": "Upload Files"
      }
    },
    "/api/v2/sessions/{session_id}/outputs.zip": {
      "get": {
        "operationId": "download_outputs_zip_api_v2_sessions__session_id__outputs_zip_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          }
        },
        "security": [
          {
            "SessionBearer": []
          }
        ],
        "summary": "Download Outputs Zip"
      }
    },
    "/api/v2/sessions/{session_id}/outputs/{filename}": {
      "get": {
        "operationId": "download_output_api_v2_sessions__session_id__outputs__filename__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "filename",
            "required": true,
            "schema": {
              "title": "Filename",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          }
        },
        "security": [
          {
            "SessionBearer": []
          }
        ],
        "summary": "Download Output"
      }
    },
    "/api/v2/sessions/{session_id}/preview-jobs": {
      "post": {
        "operationId": "create_preview_job_api_v2_sessions__session_id__preview_jobs_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewJob"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          }
        },
        "security": [
          {
            "SessionBearer": []
          }
        ],
        "summary": "Create Preview Job"
      }
    },
    "/api/v2/sessions/{session_id}/preview-jobs/{job_id}": {
      "get": {
        "operationId": "get_preview_job_api_v2_sessions__session_id__preview_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewJob"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          }
        },
        "security": [
          {
            "SessionBearer": []
          }
        ],
        "summary": "Get Preview Job"
      }
    },
    "/api/v2/sessions/{session_id}/preview-jobs/{job_id}/steps": {
      "post": {
        "operationId": "step_preview_job_api_v2_sessions__session_id__preview_jobs__job_id__steps_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewJob"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          }
        },
        "security": [
          {
            "SessionBearer": []
          }
        ],
        "summary": "Step Preview Job"
      }
    }
  }
}